app/auto_software_dev/requirements.txt
llama-cpp-python
gradio
huggingface_hub
